// frontend-next/types/generated/models/CustomTechAssessmentRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 自定义技能测验请求模型
 */
export type CustomTechAssessmentRequest = {
    /**
     * 自定义技术栈名称
     */
    technology: string;
    /**
     * 能力级别
     */
    proficiency: string;
};